renaming index to app
