+++ src/server/test/test_api.js
@@ -4,7 +4,7 @@ const chai = require('chai');
 const chai_http = require('chai-http');
 const should = chai.should();
 
-let app = require('../main/index');
+let app = require('../main/app');
 
 chai.use(chai_http);
